splitting Highcharts integration to multiple files
