+++ verticapy/plotting/_highcharts/highchart.py
@@ -22,21 +22,22 @@ permissions and limitations under the License.
 # Standard Python Modules
 from collections.abc import Iterable
 
-# High Chart
-try:
-    from vertica_highcharts import Highchart, Highstock
-except:
-    raise ImportError(
-        "The vertica_highcharts module doesn't seem to be installed in your environment.\n"
-        "To be able to use this method, you'll have to install it.\n[Tips] Run: "
-        "'pip3 install python-highcharts' in your terminal to install the module."
-    )
-
 # VerticaPy Modules
 from verticapy.connect import current_cursor
 from verticapy.sql.read import vDataFrameSQL
 from verticapy._utils._sql import _executeSQL
 from verticapy.plotting._colors import gen_colors
+
+from verticapy.plotting._highcharts.bar import bar
+from verticapy.plotting._highcharts.boxplot import boxplot
+from verticapy.plotting._highcharts.candlestick import candlestick
+from verticapy.plotting._highcharts.drilldown_chart import drilldown_chart
+from verticapy.plotting._highcharts.heatmap import heatmap
+from verticapy.plotting._highcharts.line import line
+from verticapy.plotting._highcharts.negative_bar import negative_bar
+from verticapy.plotting._highcharts.pie import pie
+from verticapy.plotting._highcharts.scatter import scatter
+from verticapy.plotting._highcharts.spider import spider
 
 #
 ##
@@ -679,901 +680,3 @@ def hchartSQL(
         max_cardinality=100,
         stock=stock,
     )
-
-
-#####
-#####
-#####
-#####
-
-
-def bar(
-    query: str,
-    options: dict = {},
-    width: int = 600,
-    height: int = 400,
-    chart_type: str = "regular",
-):
-    is_stacked = "stacked" in chart_type
-    if chart_type == "stacked_hist":
-        chart_type = "hist"
-    if chart_type == "stacked_bar":
-        chart_type = "bar"
-    data = _executeSQL(
-        query,
-        title=(
-            "Selecting the categories and their"
-            " respective aggregations to draw the chart."
-        ),
-        method="fetchall",
-    )
-    names = [desc[0] for desc in current_cursor().description]
-    n = len(names)
-    chart = Highchart(width=width, height=height)
-    if chart_type == "hist":
-        default_options = {
-            "title": {"text": ""},
-            "chart": {"type": "column"},
-            "xAxis": {"type": "category"},
-            "legend": {"enabled": False},
-        }
-    else:
-        default_options = {
-            "title": {"text": ""},
-            "chart": {"inverted": True},
-            "xAxis": {
-                "reversed": False,
-                "title": {"enabled": True, "text": names[-2]},
-                "maxPadding": 0.05,
-                "showLastLabel": True,
-            },
-            "yAxis": {"title": {"text": names[-1]}},
-            "legend": {"enabled": False},
-        }
-    default_options["colors"] = gen_colors()
-    chart.set_dict_options(default_options)
-    columns = data_to_columns(data, n)
-    if n == 2:
-        for i in range(len(columns[0])):
-            if columns[0][i] == None:
-                columns[0][i] = "None"
-        chart.set_dict_options(
-            {
-                "xAxis": {"categories": columns[0]},
-                "yAxis": {"title": {"text": names[1]}},
-            }
-        )
-        chart.set_dict_options(
-            {"tooltip": {"headerFormat": "", "pointFormat": "{point.y}"}}
-        )
-        chart.add_data_set(columns[1], "bar", names[-1], colorByPoint=True)
-    elif n == 3:
-        all_categories = sort_classes(list(set(columns[0])))
-        all_subcategories = sort_classes(list(set(columns[1])))
-        dict_categories = {}
-        for elem in all_categories:
-            dict_categories[elem] = {}
-        for i in range(len(columns[0])):
-            dict_categories[columns[0][i]][columns[1][i]] = columns[2][i]
-        for idx, elem in enumerate(dict_categories):
-            data = []
-            for cat in all_subcategories:
-                try:
-                    data += [dict_categories[elem][cat]]
-                except:
-                    data += [None]
-            chart.add_data_set(data, "bar", name=str(elem))
-        chart.set_dict_options(
-            {
-                "xAxis": {"categories": all_subcategories},
-                "yAxis": {"title": {"text": names[2]}},
-                "legend": {"enabled": True, "title": {"text": names[0]}},
-                "plotOptions": {"bar": {"dataLabels": {"enabled": True}}},
-            }
-        )
-        if is_stacked:
-            chart.set_dict_options({"plotOptions": {"series": {"stacking": "normal"}}})
-    chart.set_dict_options(options)
-    return chart
-
-
-def boxplot(
-    data: list = [],
-    options: dict = {},
-    width: int = 600,
-    height: int = 400,
-    vdf=None,
-    columns: list = [],
-    by: str = "",
-):
-    chart = Highchart(width=width, height=height)
-    default_options = {
-        "chart": {"type": "boxplot"},
-        "title": {"text": ""},
-        "legend": {"enabled": False},
-        "xAxis": {"title": {"text": ""}},
-        "yAxis": {"title": {"text": ""}},
-    }
-    default_options["colors"] = gen_colors()
-    chart.set_dict_options(default_options)
-    aggregations = ["min", "approx_25%", "approx_50%", "approx_75%", "max"]
-    if (vdf) and not (by):
-        x = vdf.agg(func=aggregations, columns=columns).transpose().values
-        data = [x[elem] for elem in x]
-        del data[0]
-        chart.set_dict_options({"xAxis": {"categories": columns}})
-        chart.set_dict_options({"yAxis": {"title": {"text": "Observations"}}})
-        title = "Observations"
-    elif vdf:
-        categories = vdf[by].distinct()
-        data = []
-        for elem in categories:
-            data += [
-                vdf.search(f"{by} = '{elem}'", usecols=[columns[0], by])[columns[0]]
-                .agg(func=aggregations)
-                .values[columns[0]]
-            ]
-        chart.set_dict_options({"xAxis": {"categories": categories}})
-        chart.set_dict_options({"yAxis": {"title": {"text": str(columns[0])}}})
-        title = by
-    chart.add_data_set(
-        data,
-        "boxplot",
-        title,
-        tooltip={"headerFormat": "<em>{point.key}</em><br/>"},
-        colorByPoint=True,
-    )
-    chart.set_dict_options(options)
-    return chart
-
-
-def candlestick(query: str, options: dict = {}, width: int = 600, height: int = 400):
-    data = _executeSQL(
-        query,
-        title=(
-            "Selecting the categories and their "
-            "respective aggregations to draw the chart."
-        ),
-        method="fetchall",
-    )
-    names = [desc[0] for desc in current_cursor().description]
-    n = len(names)
-    chart = Highstock(width=width, height=height)
-    default_options = {
-        "rangeSelector": {"selected": 1},
-        "title": {"text": ""},
-        "yAxis": [
-            {
-                "labels": {"align": "right", "x": -3},
-                "title": {"text": ""},
-                "height": "60%",
-                "lineWidth": 2,
-            },
-            {
-                "labels": {"align": "right", "x": -3},
-                "title": {"text": ""},
-                "top": "65%",
-                "height": "35%",
-                "offset": 0,
-                "lineWidth": 2,
-            },
-        ],
-    }
-    default_options["colors"] = gen_colors()
-    chart.set_dict_options(default_options)
-    for i in range(len(data)):
-        for j in range(1, n):
-            try:
-                data[i][j] = float(data[i][j])
-            except:
-                pass
-    data1 = [[elem[0], elem[1], elem[2], elem[3], elem[4]] for elem in data]
-    data2 = [[elem[0], elem[5]] for elem in data]
-    chart.add_data_set(data1, "candlestick", name="Candlesticks")
-    chart.add_data_set(data2, "column", yAxis=1, name="Volume")
-    chart.set_dict_options(options)
-    return chart
-
-
-def drilldown_chart(
-    query: list,
-    options: dict = {},
-    width: int = 600,
-    height: int = 400,
-    chart_type: str = "column",
-):
-    data = _executeSQL(
-        query[0],
-        title=(
-            "Selecting the categories and their "
-            "respective aggregations to draw the first part of the drilldown."
-        ),
-        method="fetchall",
-    )
-    names = [desc[0] for desc in current_cursor().description]
-    chart = Highchart(width=width, height=height)
-    default_options = {
-        "chart": {"type": "column"},
-        "title": {"text": ""},
-        "subtitle": {"text": ""},
-        "xAxis": {"type": "category"},
-        "yAxis": {"title": {"text": names[1]}},
-        "legend": {"enabled": False},
-        "plotOptions": {"series": {"borderWidth": 0, "dataLabels": {"enabled": True}}},
-        "tooltip": {
-            "headerFormat": "",
-            "pointFormat": '<span style="color:{point.color}">{point.name}</span>: <b>{point.y}</b><br/>',
-        },
-    }
-    default_options["colors"] = gen_colors()
-    chart.set_dict_options(default_options)
-    if chart_type == "bar":
-        chart.set_dict_options({"chart": {"inverted": True}})
-    data_final = []
-    for elem in data:
-        try:
-            val = float(elem[1])
-        except:
-            val = elem[1]
-        key = str(elem[0])
-        data_final += [{"name": key, "y": val, "drilldown": key}]
-    chart.add_data_set(data_final, chart_type, colorByPoint=True)
-    data = _executeSQL(
-        query[1],
-        title=(
-            "Selecting the categories and their respective aggregations "
-            "to draw the second part of the drilldown."
-        ),
-        method="fetchall",
-    )
-    names = [desc[0] for desc in current_cursor().description]
-    n = len(names)
-    all_categories = list(set([elem[0] for elem in data]))
-    categories = {}
-    for elem in all_categories:
-        categories[elem] = []
-    for elem in data:
-        categories[elem[0]] += [[str(elem[1]), elem[2]]]
-    for elem in categories:
-        chart.add_drilldown_data_set(
-            categories[elem], chart_type, str(elem), name=str(elem)
-        )
-    chart.set_dict_options(options)
-    chart.add_JSsource("https://code.highcharts.com/6/modules/drilldown.js")
-    return chart
-
-
-def heatmap(
-    query: str = "",
-    data: list = [],
-    options: dict = {},
-    width: int = 600,
-    height: int = 400,
-):
-    chart = Highchart(width=width, height=height)
-    default_options = {
-        "chart": {
-            "type": "heatmap",
-            "marginTop": 40,
-            "marginBottom": 80,
-            "plotBorderWidth": 1,
-        },
-        "title": {"text": ""},
-        "legend": {},
-        "colorAxis": {"minColor": "#FFFFFF", "maxColor": gen_colors()[0]},
-        "xAxis": {"title": {"text": ""}},
-        "yAxis": {"title": {"text": ""}},
-        "tooltip": {
-            "formatter": (
-                "function () {return '<b>[' + this.series.xAxis."
-                "categories[this.point.x] + ', ' + this.series.yAxis"
-                ".categories[this.point.y] + ']</b>: ' + this.point"
-                ".value + '</b>';}"
-            )
-        },
-    }
-    default_options["colors"] = gen_colors()
-    chart.set_dict_options(default_options)
-    if query:
-        data = _executeSQL(
-            query,
-            title=(
-                "Selecting the categories and their respective "
-                "aggregations to draw the chart."
-            ),
-            method="fetchall",
-        )
-        names = [desc[0] for desc in current_cursor().description]
-        n = len(names)
-        columns = data_to_columns(data, n)
-        all_categories = list(set(columns[0]))
-        all_subcategories = list(set(columns[1]))
-        dict_categories = {}
-        for elem in all_categories:
-            dict_categories[elem] = {}
-        for i in range(len(columns[0])):
-            dict_categories[columns[0][i]][columns[1][i]] = columns[2][i]
-        data = []
-        for idx, elem in enumerate(dict_categories):
-            for idx2, cat in enumerate(all_subcategories):
-                try:
-                    data += [[idx, idx2, dict_categories[elem][cat]]]
-                except:
-                    data += [[idx, idx2, None]]
-        for i in range(len(all_categories)):
-            if all_categories[i] == None:
-                all_categories[i] = "None"
-        for i in range(len(all_subcategories)):
-            if all_subcategories[i] == None:
-                all_subcategories[i] = "None"
-        chart.set_dict_options(
-            {
-                "xAxis": {"categories": all_categories, "title": {"text": names[0]},},
-                "yAxis": {
-                    "categories": all_subcategories,
-                    "title": {"text": names[1]},
-                },
-            }
-        )
-    chart.set_options(
-        "legend",
-        {
-            "align": "right",
-            "layout": "vertical",
-            "margin": 0,
-            "verticalAlign": "top",
-            "y": 25,
-            "symbolHeight": height * 0.8 - 25,
-        },
-    )
-    chart.add_data_set(
-        data,
-        series_type="heatmap",
-        borderWidth=1,
-        dataLabels={"enabled": True, "color": "#000000"},
-    )
-    chart.set_dict_options(options)
-    return chart
-
-
-def line(
-    query: str,
-    options: dict = {},
-    width: int = 600,
-    height: int = 400,
-    chart_type: str = "line",
-    stock: bool = False,
-):
-    is_ts = True if (chart_type == "area_ts") else False
-    is_range = True if (chart_type == "area_range") else False
-    is_date = False
-    is_multi = True if ("multi" in chart_type) else False
-    if chart_type in ("area_ts", "area_range", "multi_area"):
-        chart_type = "area"
-    if chart_type == "multi_line":
-        chart_type = "line"
-    if chart_type == "multi_spline":
-        chart_type = "spline"
-    data = _executeSQL(
-        query,
-        title="Selecting the different values to draw the chart.",
-        method="fetchall",
-    )
-    names = [desc[0] for desc in current_cursor().description]
-    n = len(names)
-    if stock:
-        chart = Highstock(width=width, height=height)
-        default_options = {
-            "rangeSelector": {"selected": 0},
-            "title": {"text": ""},
-            "tooltip": {
-                "style": {"width": "200px"},
-                "valueDecimals": 4,
-                "shared": True,
-            },
-            "yAxis": {"title": {"text": ""}},
-        }
-    else:
-        chart = Highchart(width=width, height=height)
-        default_options = {
-            "title": {"text": ""},
-            "xAxis": {
-                "reversed": False,
-                "title": {"enabled": True, "text": names[0]},
-                "startOnTick": True,
-                "endOnTick": True,
-                "showLastLabel": True,
-            },
-            "yAxis": {"title": {"text": names[1] if len(names) == 2 else ""}},
-            "legend": {"enabled": False},
-            "plotOptions": {
-                "scatter": {
-                    "marker": {
-                        "radius": 5,
-                        "states": {
-                            "hover": {"enabled": True, "lineColor": "rgb(100,100,100)",}
-                        },
-                    },
-                    "states": {"hover": {"marker": {"enabled": False}}},
-                    "tooltip": {
-                        "headerFormat": "",
-                        "pointFormat": "[{point.x}, {point.y}]",
-                    },
-                }
-            },
-        }
-    default_options["colors"] = gen_colors()
-    chart.set_dict_options(default_options)
-    for i in range(len(data)):
-        if "datetime" in str(type(data[i][0])):
-            is_date = True
-        for j in range(n):
-            try:
-                data[i][j] = float(data[i][j])
-            except:
-                pass
-    if is_date:
-        chart.set_options("xAxis", {"type": "datetime", "dateTimeLabelFormats": {}})
-    if n == 2:
-        chart.add_data_set(data, chart_type, names[1])
-    elif (n >= 3) and (is_multi):
-        for i in range(1, n):
-            chart.add_data_set([elem[i] for elem in data], chart_type, name=names[i])
-        chart.set_dict_options(
-            {
-                "legend": {"enabled": True},
-                "plotOptions": {
-                    "area": {
-                        "stacking": "normal",
-                        "lineColor": "#666666",
-                        "lineWidth": 1,
-                        "marker": {"lineWidth": 1, "lineColor": "#666666"},
-                    }
-                },
-                "tooltip": {"shared": True},
-                "xAxis": {
-                    "categories": [elem[0] for elem in data],
-                    "tickmarkPlacement": "on",
-                    "title": {"enabled": False},
-                },
-            }
-        )
-        if is_date:
-            chart.set_dict_options(
-                {
-                    "xAxis": {
-                        "type": "datetime",
-                        "labels": {
-                            "formatter": (
-                                "function() {return "
-                                "Highcharts.dateFormat('%a %d %b', this.value);}"
-                            )
-                        },
-                    }
-                }
-            )
-    elif n == 3:
-        all_categories = list(set([elem[-1] for elem in data]))
-        dict_categories = {}
-        for elem in all_categories:
-            dict_categories[elem] = []
-        for i in range(len(data)):
-            dict_categories[data[i][-1]] += [[data[i][0], data[i][1]]]
-        for idx, elem in enumerate(dict_categories):
-            chart.add_data_set(dict_categories[elem], chart_type, name=str(elem))
-        chart.set_dict_options(
-            {"legend": {"enabled": True, "title": {"text": names[-1]}}}
-        )
-    elif (n == 4) and (is_range):
-        data_value = [[elem[0], elem[1]] for elem in data]
-        data_range = [[elem[0], elem[2], elem[3]] for elem in data]
-        chart.add_data_set(
-            data_value,
-            "line",
-            names[1],
-            zIndex=1,
-            marker={"fillColor": "#263133", "lineWidth": 2},
-        )
-        chart.add_data_set(
-            data_range,
-            "arearange",
-            "Range",
-            lineWidth=0,
-            linkedTo=":previous",
-            fillOpacity=0.3,
-            zIndex=0,
-        )
-    if is_range:
-        chart.set_dict_options({"tooltip": {"crosshairs": True, "shared": True}})
-    if is_ts:
-        chart.set_options(
-            "plotOptions",
-            {
-                "area": {
-                    "fillColor": {
-                        "linearGradient": {"x1": 0, "y1": 0, "x2": 0, "y2": 1},
-                        "stops": [[0, "#FFFFFF"], [1, gen_colors()[0]]],
-                    },
-                    "marker": {"radius": 2},
-                    "lineWidth": 1,
-                    "states": {"hover": {"lineWidth": 1}},
-                    "threshold": None,
-                }
-            },
-        )
-    chart.set_dict_options(options)
-    return chart
-
-
-def negative_bar(query: str, options: dict = {}, width: int = 600, height: int = 400):
-    data = _executeSQL(
-        query,
-        title="Selecting the categories and their respective aggregations to draw the chart.",
-        method="fetchall",
-    )
-    names = [desc[0] for desc in current_cursor().description]
-    n = len(names)
-    chart = Highchart(width=width, height=height)
-    columns = data_to_columns(data, n)
-    all_categories = list(set(columns[0]))
-    all_subcategories = sort_classes(list(set(columns[1])))
-    default_options = {
-        "chart": {"type": "bar"},
-        "title": {"text": ""},
-        "subtitle": {"text": ""},
-        "xAxis": [
-            {
-                "categories": all_subcategories,
-                "reversed": False,
-                "labels": {"step": 1},
-            },
-            {
-                "opposite": True,
-                "reversed": False,
-                "categories": all_subcategories,
-                "linkedTo": 0,
-                "labels": {"step": 1},
-            },
-        ],
-        "yAxis": {
-            "title": {"text": None},
-            "labels": {"formatter": "function () {return (Math.abs(this.value));}"},
-        },
-        "plotOptions": {"series": {"stacking": "normal"}},
-        "tooltip": {
-            "formatter": "function () {return '<b>"
-            + names[0]
-            + " : </b>' + this.series.name + '<br>' + '<b>"
-            + names[1]
-            + "</b> : ' + '' + this.point.category + '<br/>' + '<b>"
-            + names[2]
-            + "</b> : ' + Math.abs(this.point.y);}"
-        },
-    }
-    default_options["colors"] = gen_colors()
-    chart.set_dict_options(default_options)
-    dict_categories = {}
-    for elem in all_categories:
-        dict_categories[elem] = {}
-    for i in range(len(columns[0])):
-        dict_categories[columns[0][i]][columns[1][i]] = columns[2][i]
-    for idx, elem in enumerate(dict_categories):
-        data = []
-        for cat in all_subcategories:
-            try:
-                if idx == 0:
-                    data += [dict_categories[elem][cat]]
-                else:
-                    data += [-dict_categories[elem][cat]]
-            except:
-                data += [None]
-        chart.add_data_set(data, "bar", name=str(elem))
-    chart.set_dict_options(options)
-    return chart
-
-
-def pie(
-    query: str,
-    options: dict = {},
-    width: int = 600,
-    height: int = 400,
-    chart_type: str = "regular",
-):
-    data = _executeSQL(
-        query,
-        title="Selecting the categories and their respective aggregations to draw the chart.",
-        method="fetchall",
-    )
-    names = [desc[0] for desc in current_cursor().description]
-    n = len(names)
-    chart = Highchart(width=width, height=height)
-    default_options = {
-        "title": {"text": ""},
-        "chart": {"inverted": True},
-        "xAxis": {
-            "reversed": False,
-            "title": {"text": names[0], "enabled": True},
-            "maxPadding": 0.05,
-            "showLastLabel": True,
-        },
-        "yAxis": {"title": {"text": names[1], "enabled": True}},
-        "plotOptions": {
-            "pie": {
-                "allowPointSelect": True,
-                "cursor": "pointer",
-                "showInLegend": True,
-                "size": "110%",
-            }
-        },
-        "tooltip": {"pointFormat": str(names[1]) + ": <b>{point.y}</b>"},
-    }
-    if "3d" not in chart_type:
-        default_options["colors"] = gen_colors()
-    chart.set_dict_options(default_options)
-    if "3d" in chart_type:
-        chart.set_dict_options(
-            {"chart": {"type": "pie", "options3d": {"enabled": True, "alpha": 45}}}
-        )
-        chart.add_JSsource("https://code.highcharts.com/6/highcharts-3d.js")
-        chart_type = chart_type.replace("3d", "")
-    data_pie = []
-    for elem in data:
-        try:
-            val = float(elem[1])
-        except:
-            val = elem[1]
-        try:
-            key = float(elem[0])
-        except:
-            key = elem[0]
-            if key == None:
-                key = "None"
-        data_pie += [{"name": key, "y": val}]
-    data_pie[-1]["sliced"], data_pie[-1]["selected"] = True, True
-    chart.add_data_set(data_pie, "pie")
-    if chart_type == "half":
-        chart.set_dict_options(
-            {
-                "plotOptions": {"pie": {"startAngle": -90, "endAngle": 90}},
-                "legend": {"enabled": False},
-            }
-        )
-    elif chart_type == "donut":
-        chart.set_dict_options(
-            {
-                "chart": {"type": "pie"},
-                "plotOptions": {"pie": {"innerSize": 100, "depth": 45}},
-            }
-        )
-    chart.set_dict_options(options)
-    return chart
-
-
-def scatter(
-    query: str,
-    options: dict = {},
-    width: int = 600,
-    height: int = 400,
-    chart_type: str = "regular",
-):
-    data = _executeSQL(
-        query,
-        title="Selecting the different values to draw the chart.",
-        method="fetchall",
-    )
-    names = [desc[0] for desc in current_cursor().description]
-    n = len(names)
-    chart = Highchart(width=width, height=height)
-    default_options = {
-        "title": {"text": ""},
-        "xAxis": {
-            "reversed": False,
-            "title": {"enabled": True, "text": names[0]},
-            "startOnTick": True,
-            "endOnTick": True,
-            "showLastLabel": True,
-        },
-        "yAxis": {"title": {"text": names[1]}},
-        "legend": {"enabled": False},
-        "plotOptions": {
-            "scatter": {
-                "marker": {
-                    "radius": 5,
-                    "states": {
-                        "hover": {"enabled": True, "lineColor": "rgb(100,100,100)"}
-                    },
-                },
-                "states": {"hover": {"marker": {"enabled": False}}},
-            }
-        },
-    }
-    if chart_type != "3d":
-        default_options["colors"] = gen_colors()
-    chart.set_dict_options(default_options)
-    for i in range(len(data)):
-        for j in range(n):
-            try:
-                data[i][j] = float(data[i][j])
-            except:
-                pass
-    if n == 2:
-        chart.add_data_set(data, "scatter", name="Scatter")
-        chart.set_dict_options(
-            {
-                "tooltip": {
-                    "headerFormat": "",
-                    "pointFormat": str(names[0])
-                    + ": <b>{point.x}</b><br>"
-                    + str(names[1])
-                    + ": <b>{point.y}</b>",
-                }
-            }
-        )
-    elif (n == 3) and (chart_type not in ("bubble", "3d")):
-        all_categories = list(set([elem[-1] for elem in data]))
-        dict_categories = {}
-        for elem in all_categories:
-            dict_categories[elem] = []
-        for i in range(len(data)):
-            dict_categories[data[i][-1]] += [[data[i][0], data[i][1]]]
-        for idx, elem in enumerate(dict_categories):
-            chart.add_data_set(dict_categories[elem], "scatter", name=str(elem))
-        chart.set_dict_options(
-            {"legend": {"enabled": True, "title": {"text": names[-1]}}}
-        )
-        chart.set_dict_options(
-            {
-                "tooltip": {
-                    "pointFormat": str(names[0])
-                    + ": <b>{point.x}</b><br>"
-                    + str(names[1])
-                    + ": <b>{point.y}</b>"
-                }
-            }
-        )
-    elif (n == 3) and (chart_type == "bubble"):
-        chart.add_data_set(data, "bubble", name="Bubble")
-        chart.set_dict_options(
-            {
-                "tooltip": {
-                    "headerFormat": "",
-                    "pointFormat": str(names[0])
-                    + ": <b>{point.x}</b><br>"
-                    + str(names[1])
-                    + ": <b>{point.y}</b><br>"
-                    + str(names[2])
-                    + ": <b>{point.z}</b>",
-                },
-            }
-        )
-    elif (n == 3) and (chart_type == "3d"):
-        chart.add_data_set(data, "scatter", name="Scatter")
-        chart.set_dict_options(
-            {
-                "tooltip": {
-                    "headerFormat": "",
-                    "pointFormat": str(names[0])
-                    + ": <b>{point.x}</b><br>"
-                    + str(names[1])
-                    + ": <b>{point.y}</b><br>"
-                    + str(names[2])
-                    + ": <b>{point.z}</b>",
-                }
-            }
-        )
-        chart.set_dict_options(
-            {
-                "chart": {
-                    "renderTo": "container",
-                    "margin": 100,
-                    "type": "scatter",
-                    "options3d": {
-                        "enabled": True,
-                        "alpha": 10,
-                        "beta": 30,
-                        "depth": 400,
-                        "viewDistance": 8,
-                        "frame": {
-                            "bottom": {"size": 1, "color": "rgba(0,0,0,0.02)"},
-                            "back": {"size": 1, "color": "rgba(0,0,0,0.04)"},
-                            "side": {"size": 1, "color": "rgba(0,0,0,0.06)"},
-                        },
-                    },
-                },
-                "zAxis": {"title": {"text": names[2]}},
-            }
-        )
-        chart.add_3d_rotation()
-        chart.add_JSsource("https://code.highcharts.com/6/highcharts-3d.js")
-    elif n == 4:
-        all_categories = list(set([elem[-1] for elem in data]))
-        dict_categories = {}
-        for elem in all_categories:
-            dict_categories[elem] = []
-        for i in range(len(data)):
-            dict_categories[data[i][-1]] += [[data[i][0], data[i][1], data[i][2]]]
-        if chart_type == "3d":
-            chart_type = "scatter"
-        for idx, elem in enumerate(dict_categories):
-            chart.add_data_set(dict_categories[elem], chart_type, name=str(elem))
-        chart.set_dict_options(
-            {"legend": {"enabled": True, "title": {"text": names[-1]}}}
-        )
-        if chart_type == "scatter":
-            chart.set_dict_options(
-                {
-                    "chart": {
-                        "renderTo": "container",
-                        "margin": 100,
-                        "type": "scatter",
-                        "options3d": {
-                            "enabled": True,
-                            "alpha": 10,
-                            "beta": 30,
-                            "depth": 400,
-                            "viewDistance": 8,
-                            "frame": {
-                                "bottom": {"size": 1, "color": "rgba(0,0,0,0.02)"},
-                                "back": {"size": 1, "color": "rgba(0,0,0,0.04)"},
-                                "side": {"size": 1, "color": "rgba(0,0,0,0.06)"},
-                            },
-                        },
-                    },
-                    "zAxis": {"title": {"text": names[2]}},
-                }
-            )
-            chart.add_3d_rotation()
-            chart.add_JSsource("https://code.highcharts.com/6/highcharts-3d.js")
-        chart.set_dict_options(
-            {
-                "tooltip": {
-                    "pointFormat": str(names[0])
-                    + ": <b>{point.x}</b><br>"
-                    + str(names[1])
-                    + ": <b>{point.y}</b><br>"
-                    + str(names[2])
-                    + ": <b>{point.z}</b>"
-                }
-            }
-        )
-    chart.set_dict_options(options)
-    return chart
-
-
-def spider(query: str, options: dict = {}, width: int = 600, height: int = 400):
-    data = _executeSQL(
-        query,
-        title=(
-            "Selecting the categories and their respective "
-            "aggregations to draw the chart."
-        ),
-        method="fetchall",
-    )
-    names = [desc[0] for desc in current_cursor().description]
-    n = len(names)
-    chart = Highchart(width=width, height=height)
-    default_options = {
-        "chart": {"polar": True, "type": "line", "renderTo": "test"},
-        "title": {"text": "", "x": -80},
-        "pane": {"size": "80%"},
-        "xAxis": {"tickmarkPlacement": "on", "lineWidth": 0},
-        "yAxis": {"gridLineInterpolation": "polygon", "lineWidth": 0, "min": 0},
-        "tooltip": {
-            "shared": True,
-            "pointFormat": '<span style="color:{series.color}">{series.name}: <b>{point.y:,.0f}</b><br/>',
-        },
-        "legend": {
-            "align": "right",
-            "verticalAlign": "top",
-            "y": 70,
-            "layout": "vertical",
-        },
-    }
-    default_options["colors"] = gen_colors()
-    chart.set_dict_options(default_options)
-    columns = data_to_columns(data, n)
-    chart.set_dict_options({"xAxis": {"categories": columns[0]}})
-    for i in range(1, n):
-        chart.add_data_set(columns[i], name=names[i], pointPlacement="on")
-    chart.set_dict_options(options)
-    return chart
